requirements.txt
affine==2.3.0
anyio==3.5.0
appdirs==1.4.4
arrow==1.2.3
attrs==22.1.0
barbar==0.2.1
beautifulsoup4==4.12.2
blessed==1.19.1
Bottleneck==1.3.5
brotlipy==0.7.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
captum==0.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
cleo==2.0.1
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.1
contourpy==1.0.5
crashtest==0.4.1
croniter==1.3.7
cryptography==39.0.1
cycler==0.11.0
cytoolz==0.12.0
dask==2023.4.1
dateutils==0.6.12
deepdiff==6.3.0
distlib==0.3.6
dulwich==0.21.3
efficientnet-pytorch==0.7.1
einops==0.6.1
fastapi==0.88.0
filelock==3.9.0
Fiona==1.8.22
fonttools==4.25.0
fsspec==2023.4.0

gmpy2==2.1.2
grad-cam==1.4.6
h11==0.12.0
html5lib==1.1
idna==3.4
imagecodecs==2021.8.26
imageio==2.26.0
importlib-metadata==6.0.0
importlib-resources==5.2.0
inquirer==3.1.3
installer==0.6.0
itsdangerous==2.0.1
jaraco.classes==3.2.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.1.1
jsonschema==4.17.3
keyring==23.13.1
kiwisolver==1.4.4
kornia==0.6.12
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.7.1
locket==1.0.0
lockfile==0.12.2
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
mdurl==0.1.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==8.12.0
mpmath==1.2.1
msgpack==1.0.3
munch==2.5.0
munkres==1.1.4
networkx==2.8.4
numexpr==2.8.4
numpy==1.24.3
opencv-python==4.7.0.72
ordered-set==4.1.0
orjson==3.8.8
packaging==23.0
pandas==1.5.3
partd==1.2.0
pexpect==4.8.0
Pillow==9.4.0
pip==23.0.1
pipenv==2023.5.19
pkginfo==1.9.6
platformdirs==2.5.2
ply==3.11
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
pooch==1.4.0
pretrainedmodels==0.7.4
psutil==5.9.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyproj==2.6.1.post1
pyproject_hooks==1.0.0
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==1.9.3
pytorch-metric-learning==2.1.1
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
quantus==0.4.0
rapidfuzz==2.13.7
rasterio==1.2.10
readchar==4.0.5.dev0
requests==2.29.0
requests-toolbelt==0.9.1
rich==13.3.5
Rtree==1.0.1
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.1
segmentation-models-pytorch==0.3.1
setuptools==67.8.0
shapely==2.0.1
shellingham==1.5.0
sip==6.6.2
six==1.16.0
sniffio==1.2.0
snuggs==1.4.7
soupsieve==2.4
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
threadpoolctl==2.2.0
tifffile==2021.7.2
timm==0.4.12
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchgeo==0.4.1
torchmetrics==0.11.2
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
triton==2.0.0
trove-classifiers==2023.3.9
ttach==0.0.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
websockets==10.4
wheel==0.38.4
zipp==3.11.0
